requirements.txt
alembic==1.13.2
asarPy==1.0.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Minify==0.48
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
htmlminf==0.1.13
itsdangerous==2.2.0
Jinja2==3.1.4
jsmin==3.0.1
lesscpy==0.15.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
ply==3.11
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rcssmin==1.1.2
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
xxhash==3.4.1
